small bugfix umap legend
